refactor: rename module `extracting` to `derived_spec`
